SlackBot/requirements.txt
slackclient == 2.2.1